아이템 획득 함수 추가, 탄창 text추가

+++ Enter the Gungeon/PlayerObjects/Item/Weapon/Pad.h
@@ -43,8 +43,8 @@ public:
 
 	virtual float GetReloadTime()override { return reload; }
 	virtual float GetCurrentReloadTime()override { return reloadtick; }
-	virtual float GetCurrentBulleCount() override { return currentbulletcount; }
-	virtual float GetRemainingAmmo() override { return bulletmax; }
+	virtual int GetCurrentBulleCount() override { return currentbulletcount; }
+	virtual int GetRemainingAmmo() override { return bulletmax; }
 
 	virtual void RequestReload() override { currentbulletcount = bulletcount; }
