+++ Enter the Gungeon/PlayerObjects/Item/Weapon/PilotWeapon.h
@@ -35,8 +35,8 @@ public:
 	virtual void RequestReload() override { currentbulletcount = bulletcount; }
 	virtual void GetAmmunition() override { currentbulletmax = bulletmax; }
 
-	virtual float GetCurrentBulleCount() override { return currentbulletcount; }
-	virtual float GetRemainingAmmo() override { return bulletmax; }
+	virtual int GetCurrentBulleCount() override { return currentbulletcount; }
+	virtual int GetRemainingAmmo() override { return bulletmax; }
 
 	virtual AnimationController* GetWeaponAnimation() override { return &gun; }
